:sparkles: feat: Add MoveTrackerToLog command
This commit adds a new command called MoveTrackerToLog to the app/Console/Commands directory. The command is used to move tracker data to log files grouped by year and month. It accepts an optional --month flag to specify the month to process. The command validates the month format and retrieves the tracker data for the specified month from the trackersql table. It then counts the number of records and generates a log file name based on the month. The command creates a directory for the year if it doesn't exist and writes the log data to the file. If successful, it outputs the path of the created log file.