🏃🏻feat: Add script to move tracker data to log
This commit adds a new script, "run_tracker_log.sh", which is used to move tracker data to the log. The script runs the "tracker:move-to-log" command with the specified month parameter. It is important to note that the command should be run within a Laravel project folder and the month and year parameters should be adjusted accordingly.